extract agent functions
